huge refactoring, remove antd components, move everything into app folder and redesign backend api using routes handler instead of api routes, remove use of formidable
