+++ package.json
@@ -12,14 +12,13 @@
     "@types/node": "20.3.3",
     "@types/react": "18.2.14",
     "@types/react-dom": "18.2.6",
-    "antd": "^5.7.0",
     "autoprefixer": "10.4.14",
     "bootstrap": "^5.3.1",
     "eslint": "8.44.0",
     "eslint-config-next": "13.4.8",
-    "formidable": "^3.5.0",
     "highcharts": "^11.1.0",
     "highcharts-react-official": "^3.2.0",
+    "js-yaml": "^4.1.0",
     "next": "13.4.7",
     "postcss": "8.4.24",
     "react": "18.2.0",
